+++ src/Symfony/Component/Validator/Mapping/Loader/PropertyInfoLoader.php
@@ -16,6 +16,8 @@ use Symfony\Component\PropertyInfo\PropertyListExtractorInterface;
 use Symfony\Component\PropertyInfo\PropertyTypeExtractorInterface;
 use Symfony\Component\PropertyInfo\Type as PropertyInfoType;
 use Symfony\Component\Validator\Constraints\All;
+use Symfony\Component\Validator\Constraints\DisableAutoMapping;
+use Symfony\Component\Validator\Constraints\EnableAutoMapping;
 use Symfony\Component\Validator\Constraints\NotBlank;
 use Symfony\Component\Validator\Constraints\NotNull;
 use Symfony\Component\Validator\Constraints\Type;
@@ -28,6 +30,8 @@ use Symfony\Component\Validator\Mapping\ClassMetadata;
  */
 final class PropertyInfoLoader implements LoaderInterface
 {
+    use AutoMappingTrait;
+
     private $listExtractor;
     private $typeExtractor;
     private $accessExtractor;
@@ -47,14 +51,12 @@ final class PropertyInfoLoader implements LoaderInterface
     public function loadClassMetadata(ClassMetadata $metadata): bool
     {
         $className = $metadata->getClassName();
-        if (null !== $this->classValidatorRegexp && !preg_match($this->classValidatorRegexp, $className)) {
-            return false;
-        }
-
         if (!$properties = $this->listExtractor->getProperties($className)) {
             return false;
         }
 
+        $loaded = false;
+        $enabledForClass = $this->isAutoMappingEnabledForClass($metadata, $this->classValidatorRegexp);
         foreach ($properties as $property) {
             if (false === $this->accessExtractor->isWritable($className, $property)) {
                 continue;
@@ -69,12 +71,22 @@ final class PropertyInfoLoader implements LoaderInterface
                 continue;
             }
 
+            $enabledForProperty = $enabledForClass;
             $hasTypeConstraint = false;
             $hasNotNullConstraint = false;
             $hasNotBlankConstraint = false;
             $allConstraint = null;
             foreach ($metadata->getPropertyMetadata($property) as $propertyMetadata) {
                 foreach ($propertyMetadata->getConstraints() as $constraint) {
+                    // Enabling or disabling auto-mapping explicitly always takes precedence
+                    if ($constraint instanceof DisableAutoMapping) {
+                        continue 3;
+                    }
+
+                    if ($constraint instanceof EnableAutoMapping) {
+                        $enabledForProperty = true;
+                    }
+
                     if ($constraint instanceof Type) {
                         $hasTypeConstraint = true;
                     } elseif ($constraint instanceof NotNull) {
@@ -87,6 +99,11 @@ final class PropertyInfoLoader implements LoaderInterface
                 }
             }
 
+            if (!$enabledForProperty) {
+                continue;
+            }
+
+            $loaded = true;
             $builtinTypes = [];
             $nullable = false;
             $scalar = true;
@@ -118,7 +135,7 @@ final class PropertyInfoLoader implements LoaderInterface
             }
         }
 
-        return true;
+        return $loaded;
     }
 
     private function getTypeConstraint(string $builtinType, PropertyInfoType $type): Type